+++ crates/config/src/lib.rs
@@ -11,7 +11,7 @@ mod config;
 mod config_writer;
 mod external_plugin;
 mod fixability;
-mod jsonc;
+pub mod jsonc;
 pub mod levenshtein;
 mod rule_pack;
 mod workspace;